requirements.txt
Django
psycopg2
reportlab
django-dbbackup 
boto3
django-storages
django-unfold
django-simple-history
django-payments
django-money
django-ckeditor